+++ utility/5x3.h
@@ -31,13 +31,13 @@ extern const size_t enable_pins[CHANNEL_COUNT];
 
 // Properties
 // Property values must be long, double, bool, long[], double[], bool[], char[], ConstantString *, (ConstantString *)[]
-extern const double steps_per_position_units_default[CHANNEL_COUNT];
+extern const long steps_per_position_units_default[CHANNEL_COUNT];
 
-extern const double velocity_max_default[CHANNEL_COUNT];
+extern const long velocity_max_default[CHANNEL_COUNT];
 
-extern const double velocity_min_default[CHANNEL_COUNT];
+extern const long velocity_min_default[CHANNEL_COUNT];
 
-extern const double acceleration_max_default[CHANNEL_COUNT];
+extern const long acceleration_max_default[CHANNEL_COUNT];
 
 extern const ConstantString * const enable_polarity_default[CHANNEL_COUNT];
 
@@ -47,7 +47,7 @@ extern const bool right_switch_stop_enabled_default[CHANNEL_COUNT];
 
 extern const bool switch_soft_stop_enabled_default[CHANNEL_COUNT];
 
-extern const double home_velocity_default[CHANNEL_COUNT];
+extern const long home_velocity_default[CHANNEL_COUNT];
 
 // Parameters
